requirements.txt
fastapi
uvicorn
google-cloud-vision
google-auth
python-docx
python-multipart